Real weighting based on the number of coupling occurrences in a class to another has been implemented in the DCCVisitor. "Relational Circle" is a new mxGraph-based visualisation that draws edges of differing weight and colour brightness based on the weight of the relation between two classes that they represent. It's a RelationalClass visualisation - and the DCC metric has been amended to use this type also.

+++ src/org/lukep/javavis/metrics/MetricMeasurementRelation.java
@@ -10,10 +10,10 @@ public class MetricMeasurementRelation {
 
 	protected IGenericModelNode source;
 	protected IGenericModelNode target;
-	protected int weight;
+	protected short weight;
 	
 	public MetricMeasurementRelation(IGenericModelNode source,
-			IGenericModelNode target, int weight) {
+			IGenericModelNode target, short weight) {
 		super();
 		this.source = source;
 		this.target = target;
@@ -27,7 +27,7 @@ public class MetricMeasurementRelation {
 	public IGenericModelNode getTarget() {
 		return target;
 	}
-	public int getWeight() {
+	public short getWeight() {
 		return weight;
 	}
